pyproject.toml
[project]
name = "datekeeper"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.18.0,<4.0.0)",
    "fastapi (>=0.115.10,<0.116.0)",
    "dishka (>=1.4.2,<2.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]


[build-system]
requires = ["setuptools>=61.0.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.poetry.group.dev.dependencies]
mypy = "^1.15.0"
ruff = "^0.9.9"
pre-commit = "^4.1.0"

[tool.mypy]
ignore_missing_imports = true
follow_imports = 'skip'
warn_no_return = false

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"